Change variant used for HGVSg deletion test - AGR-3024
There are a number of variations submitted by ZFIN where the
specified reference allele sequence doesn't match the reference
genome sequence at the specified positions.  ZFIN have requested
that these be removed.  However, one of those variations is used in
the current HGVSg deletion test so the test need to be updated to
use a different variation.