Changes due to SO:0001877 name change
The SOTerm name for SO:0001877 has changed from 'lnc_RNA' to
'lncRNA'.  This was preventing GFF entries with feature type set
as 'lnc_RNA' from loading.  The transcript_types and possible_types
arrays have been updated to 'lncRNA', and the feature_type_name
variable is set to 'lncRNA' if the old form is found.  This allows
transcripts to be loaded with either form of the name.